requirements.txt
# working with solidity & smart contracts:
hexbytes==0.3.0
# multiprocessing
joblib==1.1.0
# converting PNGs
Pillow==9.2.0
# date and time
python-dateutil==2.8.2
# database
sqlalchemy==1.4.40
# converting SVGs
svglib==1.4.1
Wand==0.6.10
web3==6.5.0
